@@ -1,6 +1,6 @@
 module nsparser
 
-go 1.14
+go 1.19
 
 require (
 	github.com/buaazp/fasthttprouter v0.1.1
@@ -8,4 +8,16 @@ require (
 	github.com/valyala/bytebufferpool v1.0.0
 	github.com/valyala/fasthttp v0.0.0-20171207120941-e5f51c11919d
 	golang.org/x/net v0.0.0-20200301022130-244492dfa37a
+)
+
+require (
+	github.com/chzyer/readline v0.0.0-20180603132655-2972be24d48e // indirect
+	github.com/cpuguy83/go-md2man/v2 v2.0.2 // indirect
+	github.com/manifoldco/promptui v0.9.0 // indirect
+	github.com/russross/blackfriday/v2 v2.1.0 // indirect
+	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
+	github.com/urfave/cli v1.22.10 // indirect
+	github.com/urfave/cli/v2 v2.17.1 // indirect
+	github.com/xrash/smetrics v0.0.0-20201216005158-039620a65673 // indirect
+	golang.org/x/sys v0.0.0-20190215142949-d0b11bdaac8a // indirect
 )
